refactor: address daily quality findings for 2026-08-03
Extract agent hooks child lifecycle into its own module and move agent
hook toast UI out of the notification subscription hook. Includes the
daily quality report (score 82).
